Move createEnemyWave method to EntitySpawner and write tests for method
